refactored includes + warnings
refactored includes to prefere forward declaration instread of includes

premake:
- added tag Shark to all Shark projects
- increased warning level and added fatal warnings

+++ Shark/src/Platform/Windows/WindowsThreading.cpp
@@ -132,7 +132,7 @@ namespace Shark::Threading {
 
 	void Threading::ProcessSignal::Wait(std::chrono::milliseconds time)
 	{
-		WaitForSingleObject(m_Handle, time.count());
+		WaitForSingleObject(m_Handle, static_cast<DWORD>(time.count()));
 	}
 
 	Threading::ProcessSignal::operator bool() const
